feat: restructure the package and update documentation
