--- api.d.ts
@@ -1,3 +1,0 @@
-import API from './dist/api';
-
-export = API;
