--- api.js
@@ -1,2 +1,0 @@
-const API = require('./dist/api');
-module.exports = API;
