+++ examples/destroy.js
@@ -1,10 +1,9 @@
-let CIO = require('../track');
-// In actual use require the node module: let CIO = require('customerio-node');
+const { TrackClient, RegionUS, RegionEU } = require('../index');
+// In actual use require the node module:
+// const { TrackClient, RegionUS, RegionEU } = require('customerio-node');
 const siteId = require('./config').siteId;
 const apiKey = require('./config').apiKey;
 const customerId = require('./config').customerId;
-// In actual use, specify your specific region and require node module: const { RegionUS, RegionEU } = require('customerio-node/regions')
-const { RegionUS, RegionEU } = require('../regions');
-const cio = new CIO(siteId, apiKey, { region: RegionUS });
+const cio = new TrackClient(siteId, apiKey, { region: RegionUS });
 
 cio.destroy(customerId);
