--- regions.d.ts
@@ -1,3 +1,0 @@
-import * as Regions from './dist/regions';
-
-export = Regions;
