--- regions.js
@@ -1,2 +1,0 @@
-const Regions = require('./dist/regions');
-module.exports = Regions;
