--- track.d.ts
@@ -1,3 +1,0 @@
-import TrackClient from './dist/track';
-
-export = TrackClient;
