--- track.js
@@ -1,2 +1,0 @@
-const Track = require('./dist/track');
-module.exports = Track.default;
